+++ eras/allegra/impl/testlib/Test/Cardano/Ledger/Allegra/Imp.hs
@@ -6,22 +6,18 @@
 
 module Test.Cardano.Ledger.Allegra.Imp (spec) where
 
-import Cardano.Ledger.Allegra (AllegraEra)
 import Cardano.Ledger.Core
 import Cardano.Ledger.Shelley.Rules
 import Test.Cardano.Ledger.Allegra.ImpTest
 import Test.Cardano.Ledger.Imp.Common
-import qualified Test.Cardano.Ledger.Shelley.Imp as ShelleyImp
+import qualified Test.Cardano.Ledger.Shelley.Imp as Shelley
 
 spec ::
   ( ShelleyEraImp era
-  , EraSpecificSpec era
   , Event (EraRule "RUPD" era) ~ RupdEvent
   ) =>
   proxy era ->
   Spec
 spec era = do
-  ShelleyImp.spec era
-
-instance EraSpecificSpec AllegraEra where
-  eraSpecificSpec = ShelleyImp.shelleyEraSpecificSpec
+  Shelley.spec era
+  describe "AllegraEra Onwards" $ pure ()