+++ eras/alonzo/impl/test/Main.hs
@@ -17,10 +17,14 @@ import Test.Cardano.Ledger.Alonzo.ImpTest ()
 import Test.Cardano.Ledger.Common
 import Test.Cardano.Ledger.Core.JSON (roundTripJsonEraSpec)
 import Test.Cardano.Ledger.Era
+import qualified Test.Cardano.Ledger.Shelley.Imp as Shelley
 import Test.Cardano.Ledger.Shelley.JSON (roundTripJsonShelleyEraSpec)
 
 instance EraSpec AlonzoEra where
-  eraImpSpec = Imp.spec
+  eraImpSpec era = do
+    Shelley.shelleyEraSpecificSpec era
+    Imp.alonzoEraSpecificSpec era
+    Imp.spec era
 
 main :: IO ()
 main =